- Game chơi được, gần hoàn thiện. - Thiếu âm thanh, thời gian và một vài giao diện.
