--- Flag/app/src/main/res/layout/test_gameentry.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".activity_entry"
-    android:background="#009999"
-    tools:ignore="MissingDefaultResource">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <ImageView
-            android:id="@+id/ImageEntryTheme"
-            android:layout_width="match_parent"
-            android:layout_height="300dp"
-            android:scaleType="fitXY"
-            android:src="@drawable/transparentwm"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="0dp"
-            tools:layout_editor_absoluteY="215dp"
-            android:layout_marginTop="250dp"/>
-
-        <Button
-            android:id="@+id/ButtonEntryPG"
-            android:layout_width="250dp"
-            android:layout_height="75dp"
-            android:backgroundTint="#CCFFFF"
-            android:text="Play Game"
-            android:textColor="@color/black"
-            android:textSize="28dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="215dp"
-            android:layout_marginTop="250dp"
-            android:layout_centerHorizontal="true"/>
-
-        <Button
-            android:id="@+id/ButtonEntryHS"
-            android:layout_width="250dp"
-            android:layout_height="75dp"
-            android:backgroundTint="#CCFFFF"
-            android:text="High Score"
-            android:textColor="@color/black"
-            android:textSize="28dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="328dp"
-            android:layout_below="@+id/ButtonEntryPG"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="30dp"/>
-
-        <Button
-            android:id="@+id/ButtonEntryExit"
-            android:layout_width="250dp"
-            android:layout_height="75dp"
-            android:backgroundTint="#CCFFFF"
-            android:text="Quit"
-            android:textColor="@color/black"
-            android:textSize="28dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="440dp"
-            android:layout_below="@+id/ButtonEntryHS"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="30dp"/>
-
-    </RelativeLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
