--- Flag/app/src/main/res/layout/test_gamehighscore.xml
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#009999"
-    tools:context=".activity_highscore">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <TextView
-            android:id="@+id/TextHStitle"
-            android:layout_width="340dp"
-            android:layout_height="600dp"
-            android:background="#99CCCC"
-            android:text="HIGH SCORE"
-            android:textAlignment="center"
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="38dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="35dp"
-            tools:layout_editor_absoluteY="65dp"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="65dp"/>
-
-        <TextView
-            android:id="@+id/TextHSScore"
-            android:layout_width="wrap_content"
-            android:layout_height="280dp"
-            android:text="0"
-            android:textAlignment="center"
-            android:textColor="@color/black"
-            android:textSize="150dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="150dp"/>
-
-        <Button
-            android:id="@+id/ButtonHSPG"
-            android:layout_width="300dp"
-            android:layout_height="60dp"
-            android:backgroundTint="#CCFFFF"
-            android:text="PLAY GAME"
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="20dp"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="55dp"
-            tools:layout_editor_absoluteY="500dp"
-            android:layout_below="@+id/TextHSScore"
-            android:layout_marginTop="100dp"
-            android:layout_centerHorizontal="true"/>
-
-        <Button
-            android:id="@+id/ButtonHSExit"
-            android:layout_width="300dp"
-            android:layout_height="60dp"
-            android:backgroundTint="#CCFFFF"
-            android:text="EXIT"
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="20dp"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="55dp"
-            tools:layout_editor_absoluteY="580dp"
-            android:layout_below="@+id/ButtonHSPG"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp"/>
-    </RelativeLayout>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
