--- Flag/app/src/main/res/layout/test_gamelayout_1.xml
@@ -1,131 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#009999"
-    tools:context=".MainActivity">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:id="@+id/TextLayout1QuesID"
-            android:layout_width="300dp"
-            android:layout_height="50dp"
-            android:background="#CCFFFF"
-            android:text=""
-            android:textAlignment="center"
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="38dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="55dp"
-            tools:layout_editor_absoluteY="60dp"
-            android:layout_marginTop="60dp"
-            android:layout_centerHorizontal="true"/>
-
-        <ImageView
-            android:id="@+id/ImgLayout1Question"
-            android:layout_width="300dp"
-            android:layout_height="200dp"
-            android:src="@null"
-            android:scaleType="fitXY"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="55dp"
-            tools:layout_editor_absoluteY="126dp"
-            android:layout_marginTop="20dp"
-            android:layout_below="@+id/TextLayout1QuesID"
-            android:layout_centerHorizontal="true"/>
-
-        <Button
-            android:id="@+id/ButtonLayout1AnsA"
-            android:layout_width="250dp"
-            android:layout_height="60dp"
-            android:backgroundTint="#99CCCC"
-            android:text=""
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="425dp"
-            android:layout_below="@+id/ImgLayout1Question"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp"/>
-
-        <Button
-            android:id="@+id/ButtonLayout1AnsB"
-            android:layout_width="250dp"
-            android:layout_height="60dp"
-            android:backgroundTint="#99CCCC"
-            android:text=""
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="510dp"
-            android:layout_below="@+id/ButtonLayout1AnsA"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp"/>
-
-        <Button
-            android:id="@+id/ButtonLayout1AnsC"
-            android:layout_width="250dp"
-            android:layout_height="60dp"
-            android:backgroundTint="#99CCCC"
-            android:text=""
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="589dp"
-            android:layout_below="@+id/ButtonLayout1AnsB"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp"/>
-
-        <Button
-            android:id="@+id/ButtonLayout1AnsD"
-            android:layout_width="250dp"
-            android:layout_height="60dp"
-            android:backgroundTint="#99CCCC"
-            android:text=""
-            android:textAllCaps="true"
-            android:textColor="@color/black"
-            android:textSize="20dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_editor_absoluteX="80dp"
-            tools:layout_editor_absoluteY="347dp"
-            android:layout_below="@+id/ButtonLayout1AnsC"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp"/>
-        <TextView
-            android:id="@+id/TextLayout1Result"
-            android:layout_width="200dp"
-            android:layout_height="38dp"
-            android:layout_below="@+id/ButtonLayout1AnsD"
-            android:layout_marginTop="10dp"
-            android:background="@android:color/transparent"
-            android:text=""
-            android:textAlignment="center"
-            android:textColor="@color/black"
-            android:textSize="28dp"
-            android:textStyle="bold"
-            tools:ignore="MissingConstraints"
-            tools:layout_marginLeft="110dp"
-            tools:layout_editor_absoluteY="200dp" />
-
-    </RelativeLayout>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
